chore: Application 패키지 변경
