+++ app/src/main/java/com/apptive/braini/MainActivity.kt
@@ -1,0 +1,42 @@
+package com.apptive.braini
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.compose.rememberNavController
+import com.apptive.braini.domain.log
+import com.apptive.braini.presentation.navigation.SetupNavGraph
+import com.apptive.braini.ui.theme.LayoutPracticeTheme
+import com.kakao.sdk.common.util.Utility
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+@ExperimentalMaterialApi
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            LayoutPracticeTheme {
+                val navController = rememberNavController()
+                SetupNavGraph(navController = navController)
+            }
+        }
+    }
+}
+
+@Composable
+fun Greeting(name: String) {
+    Text(text = "Hello $name!")
+}
+
+@Preview(showBackground = true)
+@Composable
+fun DefaultPreview() {
+    LayoutPracticeTheme {
+        Greeting("Android")
+    }
+}
