+++ app/src/main/java/com/apptive/braini/domain/utils/KakaoLogin.kt
@@ -1,0 +1,9 @@
+package com.apptive.braini.domain.utils
+
+import com.kakao.sdk.auth.AuthApiClient
+import com.kakao.sdk.user.UserApiClient
+
+/** 카카오 로그인을 활용하는 클래스 */
+object KakaoLogin {
+
+}
