--- app/src/main/java/com/example/braini/_constant/UIConstant.kt
@@ -1,9 +1,0 @@
-package com.example.braini._constant
-
-import androidx.compose.ui.unit.dp
-
-object UIConstant {
-    val BUTTON_WIDTH_LONG = 250.dp
-    val BUTTON_WIDTH_SMALL = 100.dp
-    val DURATION_SPLASH_MILLI = 4000
-}
